Remove unnecessary AutocompleteCommandOptionImpl
As it delegates everything to the slash option anyway
